+++ src/services/TyreImpressionService.js
@@ -7,5 +7,15 @@ export default {
   list(filters) {
     const queryString = HelperService.buildSearchQueryString(filters)
     return axios.get(`${baseURL}${queryString}`)
+  },
+  upload(file) {
+    const formData = new FormData();
+    formData.append("file", file)
+    
+    return axios.post(`${baseURL}/upload`, formData, {
+      headers: {
+        "Content-Type": "multipart/form-data",
+      }
+    })
   }
 }