+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added tyre impression list view
 - Added VeeValidate for form validation with custom rules
 - Added validated file input form component
+- Added tyre impression image upload modal
 
 ### [Changed]